Switch to new buffer *AFTER* computing contents
Often the context where a method is called is important. For example the
line in which <return> is pressed in a issue listing is required to know
the line number and then the corresponding issue.

Switching to the new buffer before the contents are computed kills the
context, and is not cool. Compute the required string in the
context *BEFORE* the switch and then just dump the string as it is.

This also simplifies the defun `setup-buffer' which had to anticipate
arguments of type string, as well as functions.

Credits to @aaditya